rename common controller to records controller
