+++ api/webhook/validate_auth.go
@@ -25,7 +25,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
 	"github.com/chaos-mesh/chaos-mesh/api/v1alpha1"
-	"github.com/chaos-mesh/chaos-mesh/controllers/common"
+	"github.com/chaos-mesh/chaos-mesh/controllers/records"
 )
 
 var alwaysAllowedKind = []string{
@@ -90,7 +90,7 @@ func (v *AuthValidator) Handle(ctx context.Context, req admission.Request) admis
 		err := fmt.Errorf("kind %s is not support", requestKind)
 		return admission.Errored(http.StatusBadRequest, err)
 	}
-	chaos := kind.Chaos.DeepCopyObject().(common.InnerObjectWithSelector)
+	chaos := kind.Chaos.DeepCopyObject().(records.InnerObjectWithSelector)
 	if chaos == nil {
 		err := fmt.Errorf("kind %s is not support", requestKind)
 		return admission.Errored(http.StatusBadRequest, err)
