+++ controllers/chaosimpl/httpchaos/impl.go
@@ -29,7 +29,7 @@ import (
 	"github.com/chaos-mesh/chaos-mesh/api/v1alpha1"
 	"github.com/chaos-mesh/chaos-mesh/controllers/chaosimpl/httpchaos/podhttpchaosmanager"
 	"github.com/chaos-mesh/chaos-mesh/controllers/chaosimpl/iochaos/podiochaosmanager"
-	"github.com/chaos-mesh/chaos-mesh/controllers/common"
+	"github.com/chaos-mesh/chaos-mesh/controllers/records"
 	"github.com/chaos-mesh/chaos-mesh/controllers/utils/controller"
 )
 
@@ -189,8 +189,8 @@ func (impl *Impl) Recover(ctx context.Context, index int, records []*v1alpha1.Re
 	return waitForRecoverSync, nil
 }
 
-func NewImpl(c client.Client, b *podhttpchaosmanager.Builder, log logr.Logger) *common.ChaosImplPair {
-	return &common.ChaosImplPair{
+func NewImpl(c client.Client, b *podhttpchaosmanager.Builder, log logr.Logger) *records.ChaosImplPair {
+	return &records.ChaosImplPair{
 		Name:   "httpchaos",
 		Object: &v1alpha1.HTTPChaos{},
 		Impl: &Impl{
